+++ pkg/providers/papi/testdata/basic-v1/property.tf
@@ -27,16 +27,16 @@ data "akamai_property_rules_template" "rules" {
 }
 
 resource "akamai_edge_hostname" "test-edgesuite-net" {
-  contract_id   = data.akamai_contract.contract.id
-  group_id      = data.akamai_group.group.id
+  contract_id   = var.contract_id
+  group_id      = var.group_id
   ip_behavior   = "IPV6_COMPLIANCE"
   edge_hostname = "test.edgesuite.net"
 }
 
 resource "akamai_property" "test-edgesuite-net" {
   name        = "test.edgesuite.net"
-  contract_id = data.akamai_contract.contract.id
-  group_id    = data.akamai_group.group.id
+  contract_id = var.contract_id
+  group_id    = var.group_id
   product_id  = "prd_HTTP_Content_Del"
   hostnames {
     cname_from             = "test.edgesuite.net"