+++ pkg/providers/papi/testdata/basic_with_cert_provisioning_type/property.tf
@@ -27,8 +27,8 @@ data "akamai_property_rules_template" "rules" {
 }
 
 resource "akamai_edge_hostname" "test-edgesuite-net" {
-  contract_id   = data.akamai_contract.contract.id
-  group_id      = data.akamai_group.group.id
+  contract_id   = var.contract_id
+  group_id      = var.group_id
   ip_behavior   = "IPV6_COMPLIANCE"
   edge_hostname = "test.edgesuite.net"
   use_cases = jsonencode([
@@ -42,8 +42,8 @@ resource "akamai_edge_hostname" "test-edgesuite-net" {
 
 resource "akamai_property" "test-edgesuite-net" {
   name        = "test.edgesuite.net"
-  contract_id = data.akamai_contract.contract.id
-  group_id    = data.akamai_group.group.id
+  contract_id = var.contract_id
+  group_id    = var.group_id
   product_id  = "prd_HTTP_Content_Del"
   rule_format = "latest"
   hostnames {